backend/requirements.txt
# LUCID Backend Requirements

# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.4.0
langchain-openai>=1.1.0

# Vector stores and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.0

# OpenAI
openai>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Additional utilities
numpy>=1.24.0
tiktoken>=0.5.0
